Tidy up route params declarations and private access
