+++ include/boost/http.hpp
@@ -46,6 +46,5 @@
 #include <boost/http/server/cors.hpp>
 #include <boost/http/server/router.hpp>
 #include <boost/http/server/basic_router.hpp>
-#include <boost/http/server/router_types.hpp>
 
 #endif
