+++ test/unit/server/router.cpp
@@ -20,7 +20,7 @@ namespace http {
 
 struct router_test
 {
-    using params = route_params_base;
+    using params = route_params;
     using test_router = basic_router<params>;
 
     //--------------------------------------------
